refactor(desktop): centralize document payload merge semantics
